refactor: 역할에 따른 서비스 단 분리
- Read, Write, Kafka 역할을 나누기 위해 3개의 클래스로 분리

JIRA: GUC-49
resolves: #7
