+++ src/main/java/com/gucci/alarm_service/kafka/consumer/NotificationConsumer.java
@@ -1,7 +1,7 @@
 package com.gucci.alarm_service.kafka.consumer;
 
 import com.gucci.alarm_service.dto.NotificationKafkaRequest;
-import com.gucci.alarm_service.service.NotificationService;
+import com.gucci.alarm_service.service.NotificationEventHandler;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.annotation.KafkaListener;
@@ -11,12 +11,12 @@ import org.springframework.stereotype.Component;
 @RequiredArgsConstructor
 @Component
 public class NotificationConsumer {
-    private final NotificationService notificationService;
+    private final NotificationEventHandler notificationEventHandler;
 
     @KafkaListener(topics = "alarm-topic", groupId = "alarm-group", containerFactory = "kafkaListenerContainerFactory")
     public void consume(NotificationKafkaRequest message) {
         log.info("받은 Kafka 메시지 : {}", message);
 
-        notificationService.createNotification(message);
+        notificationEventHandler.handle(message);
     }
 }
